backend/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# File handling
Pillow==10.2.0

# Environment
python-decouple==3.8

# API Documentation
drf-spectacular==0.27.0

# Utils
python-dateutil==2.8.2

# Development
django-extensions==3.2.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
